refactor(music): move MusicRegisterEvent to features/music/dto package
Relocate the MusicRegisterEvent class from the root events package to a dedicated DTO subpackage within the music feature. This improves code organization by placing the event class closer to its primary consumer (MusicFeature) and clarifies its purpose as a data transfer object within the feature's domain.
