--- src/mindustrytool/events/MusicRegisterEvent.java
@@ -1,4 +1,0 @@
-package mindustrytool.events;
-
-public class MusicRegisterEvent {
-}
