+++ src/mindustrytool/features/music/dto/MusicRegisterEvent.java
@@ -1,0 +1,4 @@
+package mindustrytool.features.music.dto;
+
+public class MusicRegisterEvent {
+}
